クラスの数を2つに変更
